Методы для переводчика перенесены в компонент Interpreter в папке API
